requirements.txt
matplotlib==3.3.3
prettytable==0.7.2
scipy==1.2.0
transformers==3.2.0
torch==1.6.0
kmeans_pytorch==0.3
pytorch_metric_learning==0.9.95
elasticsearch==7.10.1
tsnecuda==2.1.0
tqdm==4.55.0
numpy==1.19.2
bcubed==1.5
beautifulsoup4==4.9.3
MulticoreTSNE==0.1
scikit_learn==0.24.1
